Fix warning from tests by extracting the app initialization into a different file
