+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "nodemon --inspect --exec babel-node ./src/app.js --watch ./src",
     "start": "babel-node ./src/app.js",
-    "test": "jest"
+    "test": "jest --runInBand --detectOpenHandles"
   },
   "repository": {
     "type": "git",
